style: modified dashboard header dropdown icon.
In the process of making many disorganized changes to my code, I had messed up the headerbar style. It had shrunk and was not aligned with the other icons in the tray. I fixed this issue while also removing the profile option in the dropdown menu. This was done because there is currently no profile page, so it would just take up useless space.

tl;dr: ugly header dropdown icon was improved and dropdown option was removed due to irrelevancy