refractor: reverted previous changes.
I had changed the registration webpage to redirect upon successful registration. Although this was working as intended, it suddenly stopped and started redirecting to the login page, which is it's previous functionality. I made no changes to the code, it simply decided not to work. This problem is beyond my comprehension.

tl;dr: application's post registration behaviour suddenly changed back to previous configuration despite me not changing any code.